Mejoras de diseno
mejorar el diseno de la visa de tabla y

+++ src/app/producto/componentes/control/control.component.ts
@@ -5,7 +5,7 @@ import { Proveedor } from 'src/app/models/proveedor';
 import { ProductoService } from 'src/app/service/producto.service';
 import { ProveedorService } from 'src/app/service/proveedor.service';
 import { TokenService } from 'src/app/service/token.service';
-
+import { ConfirmationService, MessageService } from 'primeng/api';
 @Component({
   selector: 'app-control',
   templateUrl: './control.component.html',
@@ -13,21 +13,77 @@ import { TokenService } from 'src/app/service/token.service';
 })
 export class ControlComponent implements OnInit {
 
-  productos: Producto[] = [];
+  productos: Producto[];
+  producto: Producto;
   provedores: Proveedor[] = [];
+  selectedProductos: Producto[];
   isAdmin = false;
 
+  //animaciones
+  loading: boolean = true;
+
+  //control formulario
+  submitted: boolean;
+
+  //ventada emergente
+  productDialog: boolean;
   constructor(
     private productoService: ProductoService,
-    private provedorService:ProveedorService,
+    private provedorService: ProveedorService,
     private toastr: ToastrService,
-    private tokenService: TokenService
-  ) { }
+    private tokenService: TokenService,
+    private messageService: MessageService,
+    private confirmationService: ConfirmationService
+  ) {}
 
   ngOnInit() {
     this.cargarProductos();
     this.isAdmin = this.tokenService.isAdmin();
   }
+
+  //testeo
+  //! nuevo product
+  openNew() {
+    this.producto = {};
+    this.submitted = false;
+    this.productDialog = true;
+  }
+
+  editProduct(product: Producto) {
+    this.producto = { ...product };
+    this.productDialog = true;
+  }
+
+  hideDialog() {
+    this.productDialog = false;
+    this.submitted = false;
+  }
+
+  saveProduct() {
+    this.submitted = true;
+    if (this.producto.nombre.trim()) {
+      if (this.producto.id) {
+        this.productos[this.findIndexById(this.producto.nombre)] = this.producto;
+        this.messageService.add({ severity: 'success', summary: 'Successful', detail: 'Product Updated', life: 3000 });
+      }
+      this.productos = [...this.productos];
+      this.productDialog = false;
+      this.producto = {};
+    }
+  }
+
+  findIndexById(nombre: string): number {
+    let index = -1;
+    for (let i = 0; i < this.productos.length; i++) {
+        if (this.productos[i].nombre === nombre) {
+            index = i;
+            break;
+        }
+    }
+
+    return index;
+}
+  //fin testeo
 
   cargarProductos(): void {
     this.productoService.lista().subscribe(
@@ -38,24 +94,22 @@ export class ControlComponent implements OnInit {
         console.log(err);
       }
     );
-  
   }
 
-
+  //borra funciona en prime ng
   borrar(id: number) {
-    this.productoService.delete(id).subscribe(
-      data => {
-        this.toastr.success('Producto Eliminado', 'OK', {
-          timeOut: 3000, positionClass: 'toast-top-center'
-        });
-        this.cargarProductos();
-      },
-      err => {
-        this.toastr.error(err.error.mensaje, 'Fail', {
-          timeOut: 3000, positionClass: 'toast-top-center',
-        });
+    this.confirmationService.confirm({
+      message: 'Esta seguro de eliminar el producto ?',
+      header: 'Confirm',
+      icon: 'pi pi-exclamation-triangle',
+      accept: () => {
+        this.productoService.delete(id).subscribe(
+          data => {
+            this.messageService.add({ severity: 'success', summary: 'Successful', detail: 'Products Deleted', life: 3000 });
+            this.cargarProductos();
+          });
       }
-    );
+    })
   }
